Update link to readthedocs.io to stable instead of latest
Currently, the typical links to the schematic CLI docs 
are not displaying the CLI options properly:
- https://sage-schematic.readthedocs.io/en/develop/cli_reference.html
- https://sage-schematic.readthedocs.io/en/latest/cli_reference.html

This PR changes the link to https://sage-schematic.readthedocs.io/en/stable/cli_reference.html which appears to be working properly. However, it doesn't address the cause of issue, which may eventually propagate to the `stable` docs.